requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
bitly-api-py3==0.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
googlesearch-python==1.2.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
line-bot-sdk==3.14.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
openai==1.58.1
outcome==1.3.0.post0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
